pyproject.toml
[tool.poetry]
name = "zanasonic"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8, <3.11"
scipy = "^1.8.1"
pandas = "^1.4.2"
loguru = "^0.6.0"
scikit-learn = "^1.1.1"
requests = "^2.28.0"
pyarrow = "^8.0.0"
pyjanitor = "^0.23.1"
pydantic = "^1.9.1"
strictyaml = "^1.6.1"
cleo = "^0.8.1"
kaggle = "^1.5.12"
sagemaker-experiments = "^0.1.35"
awswrangler = "^2.17.0"
sagemaker = "^2.109.0"
SQLAlchemy = "^1.4.41"
psycopg2-binary = "^2.9.3"
pg8000 = "^1.29.1"
python-dotenv = "^0.21.0"
mypy = "^0.982"


[tool.poetry.dev-dependencies]
pytest = "^7.1.2"
isort = "^5.10.1"
black = "^22.3.0"
pylint = "^2.14.2"
jupyterlab = "^3.4.3"
ipykernel = "^6.15.0"
safety = "^2.0.0"
tox = "^3.25.1"
Sphinx = "^5.1.1"
sphinx-rtd-theme = "^1.0.0"
matplotlib = "^3.5.3"
jupyter = "^1.0.0"
seaborn = "^0.11.2"
pytest-cov = "^3.0.0"

[tool.poetry.scripts]
zanasonic = "zanasonic.console.application:main"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
